+++ container-search/src/test/java/ai/vespa/llm/clients/OpenAITest.java
@@ -1,9 +1,9 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package ai.vespa.llm.clients;
 
-import ai.vespa.languagemodels.InferenceParameters;
+import ai.vespa.llm.InferenceParameters;
 import ai.vespa.llm.LlmClientConfig;
-import ai.vespa.languagemodels.completion.StringPrompt;
+import ai.vespa.llm.completion.StringPrompt;
 import com.yahoo.container.jdisc.SecretStoreProvider;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
