+++ container-search/src/test/java/ai/vespa/llm/search/RAGSearcherTest.java
@@ -1,7 +1,7 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package ai.vespa.llm.search;
 
-import ai.vespa.languagemodels.LanguageModel;
+import ai.vespa.llm.LanguageModel;
 import ai.vespa.llm.LlmSearcherConfig;
 import com.yahoo.component.ComponentId;
 import com.yahoo.component.chain.Chain;
