+++ vespajlib/src/main/java/ai/vespa/llm/LanguageModel.java
@@ -1,8 +1,8 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package ai.vespa.languagemodels;
+package ai.vespa.llm;
 
-import ai.vespa.languagemodels.completion.Completion;
-import ai.vespa.languagemodels.completion.Prompt;
+import ai.vespa.llm.completion.Completion;
+import ai.vespa.llm.completion.Prompt;
 import com.yahoo.api.annotations.Beta;
 
 import java.util.List;
