+++ vespajlib/src/main/java/ai/vespa/llm/LanguageModelException.java
@@ -1,5 +1,5 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package ai.vespa.languagemodels;
+package ai.vespa.llm;
 
 import com.yahoo.api.annotations.Beta;
