+++ vespajlib/src/main/java/ai/vespa/llm/client/openai/OpenAiClient.java
@@ -1,11 +1,11 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package ai.vespa.languagemodels.client.openai;
+package ai.vespa.llm.client.openai;
 
-import ai.vespa.languagemodels.LanguageModelException;
-import ai.vespa.languagemodels.InferenceParameters;
-import ai.vespa.languagemodels.completion.Completion;
-import ai.vespa.languagemodels.LanguageModel;
-import ai.vespa.languagemodels.completion.Prompt;
+import ai.vespa.llm.LanguageModelException;
+import ai.vespa.llm.InferenceParameters;
+import ai.vespa.llm.completion.Completion;
+import ai.vespa.llm.LanguageModel;
+import ai.vespa.llm.completion.Prompt;
 import com.yahoo.api.annotations.Beta;
 import com.yahoo.slime.ArrayTraverser;
 import com.yahoo.slime.Inspector;
