+++ vespajlib/src/main/java/ai/vespa/llm/completion/package-info.java
@@ -1,11 +1,11 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 @ExportPackage
 @PublicApi
-package ai.vespa.languagemodels.test;
-
-/**
- * Tools for writing tests when working with large language models.
- */
+package ai.vespa.llm.completion;
 
 import com.yahoo.api.annotations.PublicApi;
 import com.yahoo.osgi.annotation.ExportPackage;
+
+/**
+ * Classes for generating text completions with language models.
+ */
