+++ vespajlib/src/main/java/ai/vespa/llm/test/package-info.java
@@ -1,0 +1,11 @@
+// Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+@ExportPackage
+@PublicApi
+package ai.vespa.llm.test;
+
+/**
+ * Tools for writing tests when working with large language models.
+ */
+
+import com.yahoo.api.annotations.PublicApi;
+import com.yahoo.osgi.annotation.ExportPackage;
